Assign Adjustment Codes

Assign Adjustment Codes

Usage

assign_adjustments(code, include_keys = FALSE, ...)

Arguments

code

<chr> vector of Adjustment codes to match on

include_keys

<lgl> include keys in output; default is FALSE

...

These dots are for future extensions and must be empty.

Value

a tibble

Examples

x <- c("CO-253", "OA-23", "PI-185", "-45")

assign_adjustments(x)
#> $adj_group_desc
#> [1] "Contractual Obligations"    "Other Adjustments"         
#> [3] "Payer Initiated Reductions" NA                          
#> 
#> $adj_code_desc
#> [1] "Sequestration - reduction in federal payment"                                                                  
#> [2] "The impact of prior payer(s) adjudication including payments and/or adjustments. (Use only with Group Code OA)"
#> [3] "The rendering provider is not eligible to perform the service billed."                                         
#> [4] "Charge exceeds fee schedule/maximum allowable or contracted/legislated fee arrangement."                       
#> 

assign_adjustments(x, include_keys = TRUE)
#> $adj_group_desc
#>   match_key                match_value
#> 1        CO    Contractual Obligations
#> 2        OA          Other Adjustments
#> 3        PI Payer Initiated Reductions
#> 4      <NA>                       <NA>
#> 
#> $adj_code_desc
#>   match_key
#> 1       253
#> 2        23
#> 3       185
#> 4        45
#>                                                                                                      match_value
#> 1                                                                   Sequestration - reduction in federal payment
#> 2 The impact of prior payer(s) adjudication including payments and/or adjustments. (Use only with Group Code OA)
#> 3                                          The rendering provider is not eligible to perform the service billed.
#> 4                        Charge exceeds fee schedule/maximum allowable or contracted/legislated fee arrangement.
#> 

dplyr::tibble(code = x,
              desc = list(assign_adjustments(code)))
#> # A tibble: 4 × 2
#>   code   desc            
#>   <chr>  <list>          
#> 1 CO-253 <named list [2]>
#> 2 OA-23  <named list [2]>
#> 3 PI-185 <named list [2]>
#> 4 -45    <named list [2]>

dplyr::tibble(code = x) |>
  dplyr::mutate(desc = purrr::map(code, assign_adjustments))
#> # A tibble: 4 × 2
#>   code   desc            
#>   <chr>  <list>          
#> 1 CO-253 <named list [2]>
#> 2 OA-23  <named list [2]>
#> 3 PI-185 <named list [2]>
#> 4 -45    <named list [2]>

purrr::map_df(x, assign_adjustments)
#> # A tibble: 4 × 2
#>   adj_group_desc             adj_code_desc                                      
#>   <chr>                      <chr>                                              
#> 1 Contractual Obligations    Sequestration - reduction in federal payment       
#> 2 Other Adjustments          The impact of prior payer(s) adjudication includin…
#> 3 Payer Initiated Reductions The rendering provider is not eligible to perform …
#> 4 NA                         Charge exceeds fee schedule/maximum allowable or c…

purrr::map_df(x, assign_adjustments, include_keys = TRUE)
#> # A tibble: 4 × 2
#>   adj_group_desc$match_key $match_value               adj_code_desc$match_key
#>   <chr>                    <chr>                      <chr>                  
#> 1 CO                       Contractual Obligations    253                    
#> 2 OA                       Other Adjustments          23                     
#> 3 PI                       Payer Initiated Reductions 185                    
#> 4 NA                       NA                         45                     
#> # ℹ 1 more variable: adj_code_desc$match_value <chr>

dplyr::tibble(adj_code = x,
  purrr::map_dfr(adj_code, assign_adjustments))
#> # A tibble: 4 × 3
#>   adj_code adj_group_desc             adj_code_desc                             
#>   <chr>    <chr>                      <chr>                                     
#> 1 CO-253   Contractual Obligations    Sequestration - reduction in federal paym…
#> 2 OA-23    Other Adjustments          The impact of prior payer(s) adjudication…
#> 3 PI-185   Payer Initiated Reductions The rendering provider is not eligible to…
#> 4 -45      NA                         Charge exceeds fee schedule/maximum allow…