NCCI Add-On Code Edits

Medicare NCCI Add-On Code Edits

Usage

get_addon_edits(
  hcpcs = NULL,
  aoc_type = NULL,
  edit_type = NULL,
  current = FALSE,
  ...
)

Arguments

hcpcs

<chr> vector of 5-character HCPCS codes

aoc_type

<chr> AOC type; addon or primary

edit_type

<int> AOC edit type:

1: Only Paid if Primary is Paid. Payment Eligible if Primary Payment Eligible to Same Practitioner for Same Patient on Same DOS.

2: Some Specific Primaries. Payment Eligible if, as Determined by MAC, Primary Payment Eligible to Same Practitioner for Same Patient on Same DOS.

3: No Specific Primary Codes. Payment Eligible if, as Determined by MAC, Primary Payment Eligible to Same Practitioner for Same Patient on Same DOS.

current

<lgl> return only current edits, default is FALSE

...

These dots are for future extensions and must be empty.

Value

a tibble

Details

An AOC is a HCPCS/CPT code that describes a service that, with rare exception, is performed in conjunction with another primary service by the same practitioner. An AOC is rarely eligible for payment if it is the only procedure reported by a practitioner.

Some CPT codes are identified as Add-on Codes (AOCs), which describe a service that can only be reported in addition to a primary procedure.

For specific primary codes, AOCs shall not be reported as a supplemental service for other HCPCS/CPT codes not listed as a primary code.

AOCs permit the reporting of significant supplemental services commonly performed in addition to the primary procedure. By contrast, incidental services that are necessary to accomplish the primary procedure (e.g., lysis of adhesions in the course of an open cholecystectomy) are not separately reportable with an AOC. Similarly, complications inherent in an invasive procedure occurring during the procedure are not separately reportable. For example, control of bleeding during an invasive procedure is considered part of the procedure and is not separately reportable.

Although the AOC and primary code are normally reported for the same date of service, there are unusual circumstances where the two services may be reported for different dates of service (e.g., 99291 and 99292).

AOC Edit Types

Type 1: CPT Professional or HCPCS files define all acceptable primary codes. MACs should not allow other primary codes with Type 1 AOCs.

Type 2: CPT Professional and HCPCS files do not define any primary codes. MACs should develop their own lists of acceptable primary codes.

Type 3: CPT Professional or HCPCS files define some, but not all, acceptable primary codes. MACs should allow the listed primary codes for these AOCs but may develop their own lists of additional acceptable primary codes.

PTP Edits

In general, NCCI PTP edits do not include edits with most AOCs because edits related to the primary procedure(s) are adequate to prevent inappropriate payment for an add-on coded procedure (i.e., if an edit prevents payment of the primary procedure code, the AOC shall not be paid).

Note

Version: 2024-04-01

NCCI Add-On Code Edits

Examples

get_addon_edits(hcpcs     = c("22633", "0074T"),
                aoc_type  = "primary",
                edit_type = 1)
#> # A tibble: 5 × 9
#>   hcpcs aoc_type aoc_complements   aoc_edit_type aoc_edit_description           
#>   <chr> <chr>    <list>                    <int> <chr>                          
#> 1 22633 primary  <tibble [17 × 1]>             1 Only Paid if Primary is Paid. …
#> 2 22633 primary  <tibble [1 × 1]>              1 Only Paid if Primary is Paid. …
#> 3 22633 primary  <tibble [3 × 1]>              1 Only Paid if Primary is Paid. …
#> 4 22633 primary  <tibble [1 × 1]>              1 Only Paid if Primary is Paid. …
#> 5 22633 primary  <tibble [2 × 1]>              1 Only Paid if Primary is Paid. …
#> # ℹ 4 more variables: aoc_year_deleted <int>, aoc_edit_effective <int>,
#> #   aoc_edit_deleted <int>, aoc_notes <chr>

get_addon_edits(hcpcs = c("22630",
                          "22532",
                          "77001",
                          "88334",
                          "64727"))
#> # A tibble: 17 × 9
#>    hcpcs aoc_type aoc_complements   aoc_edit_type aoc_edit_description          
#>    <chr> <chr>    <list>                    <int> <chr>                         
#>  1 22532 primary  <tibble [15 × 1]>             1 Only Paid if Primary is Paid.…
#>  2 22532 primary  <tibble [1 × 1]>              1 Only Paid if Primary is Paid.…
#>  3 22532 primary  <tibble [3 × 1]>              1 Only Paid if Primary is Paid.…
#>  4 22532 primary  <tibble [1 × 1]>              1 Only Paid if Primary is Paid.…
#>  5 22532 primary  <tibble [1 × 1]>              3 Some Specific Primaries. Paym…
#>  6 22630 primary  <tibble [16 × 1]>             1 Only Paid if Primary is Paid.…
#>  7 22630 primary  <tibble [1 × 1]>              1 Only Paid if Primary is Paid.…
#>  8 22630 primary  <tibble [3 × 1]>              1 Only Paid if Primary is Paid.…
#>  9 22630 primary  <tibble [1 × 1]>              1 Only Paid if Primary is Paid.…
#> 10 22630 primary  <tibble [2 × 1]>              1 Only Paid if Primary is Paid.…
#> 11 64727 addon    <tibble [12 × 1]>             3 Some Specific Primaries. Paym…
#> 12 64727 addon    <tibble [1 × 1]>              3 Some Specific Primaries. Paym…
#> 13 77001 addon    <tibble [1 × 1]>              2 No Specific Primary Codes. Pa…
#> 14 77001 primary  <tibble [2 × 1]>              1 Only Paid if Primary is Paid.…
#> 15 88334 addon    <tibble [2 × 1]>              1 Only Paid if Primary is Paid.…
#> 16 88334 primary  <tibble [1 × 1]>              1 Only Paid if Primary is Paid.…
#> 17 88334 primary  <tibble [1 × 1]>              1 Only Paid if Primary is Paid.…
#> # ℹ 4 more variables: aoc_year_deleted <int>, aoc_edit_effective <int>,
#> #   aoc_edit_deleted <int>, aoc_notes <chr>

get_addon_edits(hcpcs = c("11646",
                          "39503",
                          "43116",
                          "33935"))
#> # A tibble: 3 × 9
#>   hcpcs aoc_type aoc_complements  aoc_edit_type aoc_edit_description            
#>   <chr> <chr>    <list>                   <int> <chr>                           
#> 1 33935 primary  <tibble [2 × 1]>             1 Only Paid if Primary is Paid. P…
#> 2 33935 primary  <tibble [1 × 1]>             1 Only Paid if Primary is Paid. P…
#> 3 33935 primary  <tibble [3 × 1]>             1 Only Paid if Primary is Paid. P…
#> # ℹ 4 more variables: aoc_year_deleted <int>, aoc_edit_effective <int>,
#> #   aoc_edit_deleted <int>, aoc_notes <chr>