Procedure to Procedure (PTP) Edits

National Correct Coding Initiative (NCCI) Procedure-to-Procedure (PTP) edits prevent inappropriate payment of services that should not be reported together. Each edit has a Column One and Column Two HCPCS/CPT code.

Usage

get_ptp_edits(
  hcpcs = NULL,
  ptp_type = NULL,
  ptp_edit_mod = NULL,
  current = TRUE,
  ...
)

Arguments

hcpcs

<chr> vector of 5-character HCPCS codes

ptp_type

<chr> comprehensive (Column One code) or component (Column Two code)

ptp_edit_mod

<int>

1: Allowed

0: Not Allowed

9: Not Applicable

current

<lgl> return only current edits, default is TRUE

...

These dots are for future extensions and must be empty.

Value

a tibble

Details

If a provider reports the two codes of an edit pair for the same beneficiary on the same date of service, the Column One code is eligible for payment, but the Column Two code is denied unless a clinically appropriate NCCI PTP-associated modifier is also reported.

The NCCI PTP edits and MUEs are usually updated at least quarterly.

NCCI Modifiers

Modifiers that may be used under appropriate clinical circumstances to bypass an NCCI PTP edit include:

Anatomic: E1-E4, FA, F1-F9, TA, T1-T9, LT, RT, LC, LD, RC, LM, RI

Global Surgery: 24, 25, 57, 58, 78, 79

Other: 27, 59, 91, XE, XS, XP, XU

It’s very important that NCCI PTP-associated modifiers only be used when appropriate. In general, these circumstances relate to separate patient encounters, separate anatomic sites, or separate specimens. (See subsequent discussion of modifiers in this section.)

Most edits involving paired organs or structures (e.g., eyes, ears, extremities, lungs, kidneys) have NCCI PTP modifier indicators of 1 because the two codes of the code pair edit may be reported if performed on the contralateral organs or structures.

Most of these code pairs should not be reported with NCCI PTP-associated modifiers when performed on the ipsilateral organ or structure unless there is a specific coding rationale to bypass the edit.

The presence of an NCCI PTP edit indicates that the two codes generally can’t be reported together unless the two corresponding procedures are performed at two separate patient encounters or two separate anatomic locations.

Similarly, if the two corresponding procedures are performed at the same patient encounter and in contiguous structures in the same organ or anatomic region, NCCI PTP-associated modifiers generally shouldn’t be used.

Modifier 59 (MLN 1783722) may be used only if no other appropriate modifier describes the service. The article provides more information on the appropriate use of the 59 modifier.

Some Column One/Column Two correct coding edits would ever warrant the use of any of the modifiers associated with the NCCI PTP edits. These code pairs are assigned a correct coding modifier indicator (CCMI) of 0.

Claim line edits allow use of NCCI PTP-associated Modifier 91 to bypass them if one or more of the individual laboratory tests are repeated on the same date of service.

The repeat testing must be medically reasonable and necessary and can’t be performed to "confirm initial results; due to testing problems with specimens and equipment or for any other reason when a normal, one-time, reportable result is all that is required."

Procedure to Procedure (PTP) Edits

Examples

get_ptp_edits(hcpcs = c("39503", "43116", "33935", "11646"))
#> # A tibble: 56 × 6
#>    hcpcs ptp_type      ptp_complements   ptp_edit_mod ptp_edit_mod_desc
#>    <chr> <chr>         <list>                   <int> <chr>            
#>  1 11646 comprehensive <tibble [1 × 1]>             0 Not Allowed      
#>  2 11646 comprehensive <tibble [25 × 1]>            0 Not Allowed      
#>  3 11646 comprehensive <tibble [46 × 1]>            1 Allowed          
#>  4 11646 comprehensive <tibble [24 × 1]>            1 Allowed          
#>  5 11646 comprehensive <tibble [63 × 1]>            1 Allowed          
#>  6 11646 comprehensive <tibble [4 × 1]>             1 Allowed          
#>  7 11646 comprehensive <tibble [11 × 1]>            0 Not Allowed      
#>  8 11646 comprehensive <tibble [26 × 1]>            0 Not Allowed      
#>  9 11646 comprehensive <tibble [1 × 1]>             1 Allowed          
#> 10 11646 comprehensive <tibble [3 × 1]>             0 Not Allowed      
#> 11 11646 component     <tibble [26 × 1]>            1 Allowed          
#> 12 11646 component     <tibble [9 × 1]>             1 Allowed          
#> 13 11646 component     <tibble [2 × 1]>             1 Allowed          
#> 14 11646 component     <tibble [3 × 1]>             1 Allowed          
#> 15 11646 component     <tibble [4 × 1]>             1 Allowed          
#> 16 11646 component     <tibble [1 × 1]>             1 Allowed          
#> 17 33935 comprehensive <tibble [27 × 1]>            0 Not Allowed      
#> 18 33935 comprehensive <tibble [50 × 1]>            1 Allowed          
#> 19 33935 comprehensive <tibble [63 × 1]>            1 Allowed          
#> 20 33935 comprehensive <tibble [67 × 1]>            1 Allowed          
#> # ℹ 36 more rows
#> # ℹ 1 more variable: ptp_edit_rationale <chr>

get_ptp_edits(hcpcs        = "43116",
              ptp_type     = "component",
              ptp_edit_mod = 0)
#> # A tibble: 2 × 6
#>   hcpcs ptp_type  ptp_complements  ptp_edit_mod ptp_edit_mod_desc
#>   <chr> <chr>     <list>                  <int> <chr>            
#> 1 43116 component <tibble [3 × 1]>            0 Not Allowed      
#> 2 43116 component <tibble [1 × 1]>            0 Not Allowed      
#> # ℹ 1 more variable: ptp_edit_rationale <chr>