Validate HCPCS Level I (CPT) Category II Code

A valid CPT Category II code is:

Usage

is_cpt_category_II(hcpcs_code)

Source

AMA Category II Codes

Arguments

hcpcs_code

<chr> string

Value

<lgl> TRUE if valid, otherwise FALSE

Details

5 characters long

Begins with any 4 digits, [0-9]{4}, and

Ends with an [F]

Category II codes are supplemental tracking codes that are used to track services on claims for performance measurement and are not billing codes.

These codes are intended to facilitate data collection about quality of care by coding certain services and/or test results that support performance measures and that have been agreed upon as contributing to good patient care.

Some codes in this category may relate to compliance by the health care professional with state or federal law. The use of these codes is optional. The codes are not required for correct coding and may not be used as a substitute for Category I codes.

Services/procedures or test results described in this category make use of alpha characters as the 5th character in the string (i.e., 4 digits followed by an alpha character).

These digits are not intended to reflect the placement of the code in the regular (Category I) part of the CPT code set.

Also, these codes describe components that are typically included in an evaluation and management service or test results that are part of the laboratory test/procedure. Consequently, they do not have a relative value associated with them.

Examples

x <- c('T1503', 'G0478', '81301', '69641', '0583F', '0779T', NA)

is_cpt_category_II(x)
#> [1] FALSE FALSE FALSE FALSE  TRUE FALSE    NA

x[which(is_cpt_category_II(x))]
#> [1] "0583F"