Search Adjustment Codes

CARC and RARC Codes

Usage

search_adjustments(type = c("group", "carc", "rarc"), ...)

Arguments

type

<chr> type of Adjustment code; group (default), carc, rarc

...

These dots are for future extensions and must be empty.

Value

a tibble

Details

Claim Adjustment Reason Codes:

X12 External Code Source 139

These codes describe why a claim or service line was paid differently than it was billed and generally assign responsibility for the adjustment amounts.

The Claim Adjustment Group Codes are internal to the X12 standard. The format is always two alpha characters:

CO: Contractual Obligations

CR: Corrections and Reversals

OA: Other Adjustments

PI: Payer Initiated Reductions

PR: Patient Responsibility

Remittance Advice Remark Codes:

X12 External Code Source 411

Remittance Advice Remark Codes (RARCs) are used to provide additional explanation for an adjustment already described by a Claim Adjustment Reason Code (CARC) or to convey information about remittance processing.

There are two types of RARCs: Supplemental and Informational:

The majority of RARCs are supplemental and, as such, are generally referred to as RARCs without further distinction. Supplemental RARCs provide additional explanation for an adjustment already described by a CARC.

The second type is informational; these are all prefaced with Alert: and are referred to as Alerts. They are used to convey information about remittance processing and are never related to a specific adjustment or CARC.

Examples

search_adjustments("group")
#> # A tibble: 5 × 2
#>   code  description               
#>   <chr> <chr>                     
#> 1 CO    Contractual Obligations   
#> 2 CR    Corrections and Reversals 
#> 3 OA    Other Adjustments         
#> 4 PI    Payer Initiated Reductions
#> 5 PR    Patient Responsibility    

search_adjustments("carc")
#> # A tibble: 405 × 7
#>    code  description             usage notes start_date last_modified end_date  
#>    <chr> <chr>                   <chr> <chr> <date>     <date>        <date>    
#>  1 1     Deductible Amount       NA    NA    1995-01-01 NA            NA        
#>  2 2     Coinsurance Amount      NA    NA    1995-01-01 NA            NA        
#>  3 3     Co-payment Amount       NA    NA    1995-01-01 NA            NA        
#>  4 4     The procedure code is … Refe… NA    1995-01-01 2020-03-01    NA        
#>  5 5     The procedure code/typ… Refe… NA    1995-01-01 2018-03-01    NA        
#>  6 6     The procedure/revenue … Refe… NA    1995-01-01 2017-07-01    NA        
#>  7 7     The procedure/revenue … Refe… NA    1995-01-01 2017-07-01    NA        
#>  8 8     The procedure code is … Refe… NA    1995-01-01 2017-07-01    NA        
#>  9 9     The diagnosis is incon… Refe… NA    1995-01-01 2017-07-01    NA        
#> 10 10    The diagnosis is incon… Refe… NA    1995-01-01 2017-07-01    NA        
#> 11 11    The diagnosis is incon… Refe… NA    1995-01-01 2017-07-01    NA        
#> 12 12    The diagnosis is incon… Refe… NA    1995-01-01 2017-07-01    NA        
#> 13 13    The date of death prec… NA    NA    1995-01-01 NA            NA        
#> 14 14    The date of birth foll… NA    NA    1995-01-01 NA            NA        
#> 15 15    The authorization numb… NA    NA    1995-01-01 2017-11-01    2018-05-01
#> 16 16    Claim/service lacks in… Do n… NA    1995-01-01 2018-03-01    NA        
#> 17 17    Requested information … NA    NA    1995-01-01 2008-09-21    2009-07-01
#> 18 18    Exact duplicate claim/… NA    NA    1995-01-01 2013-06-02    NA        
#> 19 19    This is a work-related… NA    NA    1995-01-01 2007-09-30    NA        
#> 20 20    This injury/illness is… NA    NA    1995-01-01 2007-09-30    NA        
#> # ℹ 385 more rows

search_adjustments("rarc")
#> # A tibble: 1,171 × 5
#>    code  description                              notes start_date last_modified
#>    <chr> <chr>                                    <chr> <date>     <date>       
#>  1 M1    X-ray not taken within the past 12 mont… NA    1997-01-01 NA           
#>  2 M2    Not paid separately when the patient is… NA    1997-01-01 NA           
#>  3 M3    Equipment is the same or similar to equ… NA    1997-01-01 NA           
#>  4 M4    Alert: This is the last monthly install… (Mod… 1997-01-01 2007-04-01   
#>  5 M5    Monthly rental payments can continue un… NA    1997-01-01 NA           
#>  6 M6    Alert: You must furnish and service thi… (Mod… 1997-01-01 2009-03-01   
#>  7 M7    No rental payments after the item is pu… (Mod… 1997-01-01 2016-11-01   
#>  8 M8    We do not accept blood gas tests result… NA    1997-01-01 NA           
#>  9 M9    Alert: This is the tenth rental month. … (Mod… 1997-01-01 2007-04-01   
#> 10 M10   Equipment purchases are limited to the … NA    1997-01-01 NA           
#> 11 M11   DME, orthotics and prosthetics must be … NA    1997-01-01 NA           
#> 12 M12   Diagnostic tests performed by a physici… NA    1997-01-01 NA           
#> 13 M13   Only one initial visit is covered per s… (Mod… 1997-01-01 2007-06-30   
#> 14 M14   No separate payment for an injection ad… NA    1997-01-01 NA           
#> 15 M15   Separately billed services/tests have b… NA    1997-01-01 NA           
#> 16 M16   Alert: Please see our web site, mailing… (Rea… 1997-01-01 2007-04-01   
#> 17 M17   Alert: Payment approved as you did not … (Mod… 1997-01-01 2007-04-01   
#> 18 M18   Certain services may be approved for ho… (Mod… 1997-01-01 2003-06-30   
#> 19 M19   Missing oxygen certification/re-certifi… (Mod… 1997-01-01 2003-02-28   
#> 20 M20   Missing/incomplete/invalid HCPCS.        (Mod… 1997-01-01 2003-02-28   
#> # ℹ 1,151 more rows