NCCI Medically Unlikely Edits

Medicare NCCI Medically Unlikely Edits (MUEs)

Usage

search_mue(hcpcs = NULL, service = NULL, mai = NULL, ...)

Arguments

hcpcs

vector of HCPCS codes

service

vector of HCPCS codes

mai

MUE adjudication indicator; 1, 2, 3

...

Empty

Value

a tibble

Details

National Correct Coding Initiative (NCCI) Medically Unlikely Edits (MUEs) are used by the Medicare Administrative Contractors (MACs), to reduce improper payments for Part B claims.

An MUE is the maximum Units of Service (UOS) reported for a HCPCS/CPT code on the vast majority of appropriately reported claims by the same provider/supplier for the same beneficiary on the same date of service.

Not all HCPCS/CPT codes have an MUE. Although CMS publishes most MUE values on its website, other MUE values are confidential. Confidential MUE values are not releasable. The confidential status of MUEs is subject to change.

Quarterly Version Update Changes

CMS posts changes to each of its NCCI MUE published edit files on a quarterly basis. This includes additions, deletions, and revisions to published MUEs for Practitioner Services, Outpatient Hospital Services, and DME Supplier Services.

CMS began introducing date of service (DOS) MUEs. Over time CMS will convert many, but not all, MUEs to DOS MUEs. MUEs are adjudicated either as claim line edits or DOS edits. If the MUE is adjudicated as a claim line edit, the UOS on each claim line are compared to the MUE value for the HCPCS/CPT code on that claim line. If the UOS exceed the MUE value, all UOS on that claim line are denied.

If the MUE is adjudicated as a DOS MUE, all UOS on each claim line for the same date of service for the same HCPCS/CPT code are summed, and the sum is compared to the MUE value. If the summed UOS exceed the MUE value, all UOS for the HCPCS/CPT code for that date of service are denied. Denials due to claim line MUEs or DOS MUEs may be appealed to the local claims processing contractor. DOS MUEs are used for HCPCS/CPT codes where it would be extremely unlikely that more UOS than the MUE value would ever be performed on the same date of service for the same patient.

MUE Adjudication Indicator (MAI)

An MAI of 1 indicates that the edit is a claim line MUE. An MAI of 2 or 3 indicates that the edit is a DOS MUE.

If a HCPCS/CPT code has an MUE that is adjudicated as a claim line edit, appropriate use of CPT modifiers (i.e., 59 or XE, XP, XS, XU; 76, 77, 91, anatomic) may be used to report the same HCPCS/CPT code on separate lines of a claim.

MUEs for HCPCS codes with an MAI of 2 are absolute date of service edits. These are "per day edits based on policy." These have been rigorously reviewed and vetted within CMS and obtain this MAI designation because UOS on the same date of service (DOS) in excess of the MUE value would be considered impossible because it was contrary to statute, regulation, or subregulatory guidance.

MUEs for HCPCS codes with an MAI of 3 are "per day edits based on clinical benchmarks." These are based on criteria (e.g., nature of service, prescribing information) combined with data such that it would be possible but medically highly unlikely that higher values would represent correctly reported medically necessary services

MUE Link

Examples

search_mue(hcpcs = c("39503", "43116", "33935", "11646"))
#> # A tibble: 8 × 6
#>   hcpcs   mue   mai adjudication                   rationale        service_type
#>   <chr> <int> <int> <chr>                          <chr>            <chr>       
#> 1 11646     2     3 Date of Service Edit: Clinical Clinical: Data   Practitioner
#> 2 33935     1     2 Date of Service Edit: Policy   Anatomic Consid… Practitioner
#> 3 39503     1     2 Date of Service Edit: Policy   Anatomic Consid… Practitioner
#> 4 43116     1     2 Date of Service Edit: Policy   Anatomic Consid… Practitioner
#> 5 11646     2     3 Date of Service Edit: Clinical Clinical: Data   Outpatient …
#> 6 33935     1     2 Date of Service Edit: Policy   Anatomic Consid… Outpatient …
#> 7 39503     1     2 Date of Service Edit: Policy   Anatomic Consid… Outpatient …
#> 8 43116     1     2 Date of Service Edit: Policy   Anatomic Consid… Outpatient …