NCCI Edits: Medically Unlikely Edits (MUEs)

National Correct Coding Initiative (NCCI) Medically Unlikely Edits (MUEs) are used by Medicare Administrative Contractors (MACs) to reduce improper payments for Part B claims.

Usage

search_mues(
  hcpcs_code = NULL,
  mue_service = c("Practitioner", "Outpatient", "DME"),
  mue_mai = NULL,
  ...
)

Arguments

hcpcs_code

<chr> vector of 5-character HCPCS codes

mue_service

<chr> Practitioner, Outpatient, or DME; default is Practitioner

mue_mai

<int> MUE adjudication indicator; 1, 2, or 3

...

These dots are for future extensions and must be empty.

Value

a tibble

Details

An MUE describes the maximum Units of Service (UOS) reported for a HCPCS code on the vast majority of appropriately reported claims by the same provider/supplier for the same beneficiary on the same date of service.

Not all HCPCS codes have an MUE. Although CMS publishes most MUE values, other MUE values are confidential. Confidential MUE values are not releasable. The confidential status of MUEs is subject to change.

MUE Edit Types

MUEs are adjudicated as either a Claim Line edit or Date of Service edit, the latter of which has two subclasses: Policy and Clinical.

If the MUE is adjudicated as a Claim Line edit, the HCPCS code's UOS on each claim line are compared to that code's MUE value. If the UOS exceeds the MUE value, all UOS on that claim line are denied.

DOS MUEs are used for HCPCS codes where it would be extremely unlikely that more UOS than the MUE value would ever be performed on the same date of service for the same patient.

If the MUE is adjudicated as a DOS MUE, the sum of all UOS on each claim line for the same date of service for the same HCPCS code is compared to the MUE value. If the summed UOS exceeds the MUE value, all UOS for the HCPCS code for that date of service are denied.

MUE Adjudication Indicator

An MAI of 1 indicates that the edit is a claim line MUE. An MAI of 2 or 3 indicates that the edit is a DOS MUE.

If a HCPCS code has an MUE that is adjudicated as a claim line edit, appropriate use of modifiers (i.e., 59 or XE, XP, XS, XU; 76, 77, 91, anatomic) may be used to report the same HCPCS code on separate lines of a claim.

MUEs for HCPCS codes with an MAI of 2 are absolute date of service edits. These are "per day edits based on policy." They obtain this MAI designation because UOS on the same date of service (DOS) in excess of the MUE value would contradict statute, regulation, or subregulatory guidance.

MUEs with an MAI of 3 are "per day edits based on clinical benchmarks." These are based on criteria (e.g., nature of service, prescribing information) and historical data such that it would be possible but medically highly unlikely that higher values would represent correctly reported medically necessary services.

Quarterly Updates

CMS posts changes to each NCCI MUE edit file on a quarterly basis. This includes additions, deletions, and revisions to published MUEs for Practitioner Services, Outpatient Hospital Services, and DME Supplier Services.

NCCI Medically Unlikely Edits

See also

Examples

search_mues(hcpcs_code = c("39503", "43116", "33935", "11646"))
#> # A tibble: 4 × 5
#>   hcpcs_code mue_units mue_mai mue_rationale          mue_service 
#>   <chr>          <int>   <int> <chr>                  <chr>       
#> 1 11646              2       3 Clinical: Data         Practitioner
#> 2 33935              1       2 Anatomic Consideration Practitioner
#> 3 39503              1       2 Anatomic Consideration Practitioner
#> 4 43116              1       2 Anatomic Consideration Practitioner