Place of Service (POS) Codes

Place of Service Codes are two-digit codes placed on health care professional claims to indicate the setting in which a service was provided. The Centers for Medicare & Medicaid Services (CMS) maintain POS codes used throughout the health care industry.

Usage

search_pos(pos_code = NULL, pos_type = NULL, ...)

Arguments

pos_code

<chr> vector of 2-character Place of Service codes; default is NULL

pos_type

<chr> Place of Service type, one of Facility or Non-Facility; default is NULL

...

These dots are for future extensions and must be empty.

Value

a tibble

Details

This code set is required for use in the implementation guide adopted as the national standard for electronic transmission of professional health care claims under the provisions of the Health Insurance Portability and Accountability Act of 1996 (HIPAA).

HIPAA directed the Secretary of HHS to adopt national standards for electronic transactions. These standard transactions require all health plans and providers to use standard code sets to populate data elements in each transaction.

The Transaction and Code Set Rule adopted the ASC X12N-837 Health Care Claim: Professional, volumes 1 and 2, version 4010, as the standard for electronic submission of professional claims.

This standard names the POS code set currently maintained by CMS as the code set to be used for describing sites of service in such claims. POS information is often needed to determine the acceptability of direct billing of Medicare, Medicaid and private insurance services provided by a given provider.

Code Structure

The first digit denotes service category and the second specifies location or service type.

POS code 11 represents an Office for direct patient services.

For instance, code 11 designates an Office, encompassing physician offices, clinics, group practices, and standalone facilities providing direct patient services.

Conversely, place of service code 22 denotes an On Campus-Outpatient Hospital, covering services in a hospital-based outpatient department where the patient is admitted as an outpatient.

Structure: First digit for category, second for location/type

Choose POS based on the majority of services in a specific encounter

Modify POS codes if the location changes during a single encounter

Location Types

Based on the location of service, POS codes can be grouped into four categories: Facility, Non-Facility, Telehealth, and Other.

Facility Codes

Facility POS codes are used to indicate services provided in a facility setting such as hospitals, nursing homes, or skilled nursing facilities. These include:

Urgent Care Facility (20)

Inpatient Hospital (21)

Outpatient Hospital (22)

Emergency Room-Hospital (23)

Ambulatory Surgical Center (24)

Skilled Nursing Facility (31)

Hospice Facility (32)

These codes indicate that the services were provided in a facility that is owned and operated by a healthcare provider.

Non-Facility Codes

Non-facility POS codes are used for services provided in non-facility settings such as physician offices or independent clinics. These include:

School (03)

Office (11)

Home (12)

Independent Clinic (49)

The above codes indicate that the services were provided in a setting not owned or operated by a healthcare provider.

Telehealth POS Codes

Telehealth place-of-service codes are used to indicate services provided through telecommunication technology. These include:

Telehealth (02)

Store and Forward Telemedicine Services (18)

This category of POS codes was introduced due to the increasing use of telehealth services in healthcare, and to differentiate them from traditional in-person services.

Other POS Codes

In addition to facility-specific POS codes, there are other codes that play a vital role in accurately describing healthcare encounters. These codes are designed for specific scenarios, such as visits to retail clinics, public health clinics, or rural health clinics. Each code serves a distinct purpose in the healthcare landscape.

Homeless Shelter (04)

Retail Clinic (17)

Rural Health Clinic (72)

These codes, including Retail Clinic (17), Home Shelter (04), and Rural Health Clinic (72), cater to unique healthcare settings, ensuring comprehensive coverage in billing and reporting. Understanding and applying these codes appropriately contributes to the precision of healthcare documentation and facilitates effective reimbursement processes.

See also

Examples

search_pos(pos_code = c("11", "21"))
#> # A tibble: 2 × 4
#>   pos_code pos_type     pos_name           pos_description                      
#>   <chr>    <fct>        <chr>              <chr>                                
#> 1 11       Non-Facility Office             Location, other than a hospital, ski…
#> 2 21       Facility     Inpatient Hospital A facility, other than psychiatric, …

search_pos(pos_type = "Facility")
#> # A tibble: 18 × 4
#>    pos_code pos_type pos_name                                    pos_description
#>    <chr>    <fct>    <chr>                                       <chr>          
#>  1 02       Facility Telehealth Provided Other than in Patient’… The location w…
#>  2 06       Facility Indian Health Service Provider-based Facil… A facility or …
#>  3 08       Facility Tribal 638 Provider-based Facility          A facility or …
#>  4 19       Facility Off Campus-Outpatient Hospital              A portion of a…
#>  5 21       Facility Inpatient Hospital                          A facility, ot…
#>  6 22       Facility On Campus-Outpatient Hospital               A portion of a…
#>  7 23       Facility Emergency Room-Hospital                     A portion of a…
#>  8 24       Facility Ambulatory Surgical Center                  A freestanding…
#>  9 26       Facility Military Treatment Facility                 A medical faci…
#> 10 31       Facility Skilled Nursing Facility                    A facility whi…
#> 11 34       Facility Hospice                                     A facility, ot…
#> 12 41       Facility Ambulance-Land                              A land vehicle…
#> 13 42       Facility Ambulance-Air or Water                      An air or wate…
#> 14 51       Facility Inpatient Psychiatric Facility              A facility tha…
#> 15 52       Facility Psychiatric Facility-Partial Hospitalizati… A facility for…
#> 16 53       Facility Community Mental Health Center              A facility tha…
#> 17 56       Facility Psychiatric Residential Treatment Center    A facility or …
#> 18 61       Facility Comprehensive Inpatient Rehabilitation Fac… A facility tha…