NCCI Procedure to Procedure (PTP) Edits

National Correct Coding Initiative (NCCI) Procedure-to-Procedure (PTP) edits prevent inappropriate payment of services that should not be reported together. Each edit has a Column One (Comprehensive) and Column Two (Component) HCPCS code.

Usage

search_ptp(
  hcpcs_code = NULL,
  ptp_service = c("Practitioner", "Outpatient"),
  ptp_type = NULL,
  ptp_mod = NULL,
  unnest = FALSE,
  ...
)

Arguments

hcpcs_code

<chr> vector of 5-character HCPCS codes

ptp_service

<chr> Practitioner or Outpatient; default is Practitioner

ptp_type

<chr> Comprehensive (Column One) or Component (Column Two)

ptp_mod

<int>

1: Allowed: The code-pair is allowed with an NCCI PTP-associated modifier.

0: Not Allowed: An NCCI PTP-associated modifier is not allowed and will not bypass the edit.

9: Not Applicable: The use of NCCI PTP-associated modifiers is not specified.

unnest

<lgl> Unnest the ptp_complements column, default is FALSE

...

These dots are for future extensions and must be empty.

Value

a tibble

Details

If a provider reports the two codes of an edit pair for the same beneficiary on the same date of service, the Column One code is eligible for payment, but the Column Two code is denied unless a clinically appropriate NCCI PTP-associated modifier is also reported.

The NCCI PTP edits and MUEs are usually updated at least quarterly.

PTP Modifiers

Modifiers that may be used under appropriate clinical circumstances to bypass an NCCI PTP edit include:

Anatomic: E1-E4, FA, F1-F9, TA, T1-T9, LT, RT, LC, LD, RC, LM, RI

Global Surgery: 24, 25, 57, 58, 78, 79

Other: 27, 59, 91, XE, XS, XP, XU

It’s very important that NCCI PTP-associated modifiers only be used when appropriate. In general, these circumstances relate to separate patient encounters, separate anatomic sites, or separate specimens.

Most edits involving paired organs or structures (e.g., eyes, ears, extremities, lungs, kidneys) have NCCI PTP modifier indicators of 1 because the two codes of the code-pair edit may be reported if performed on the contralateral organs or structures.

Most of these code-pairs should not be reported with NCCI PTP-associated modifiers when performed on the ipsilateral organ or structure unless there is a specific coding rationale to bypass the edit.

The presence of an NCCI PTP edit indicates that the two codes generally can’t be reported together unless the two corresponding procedures are performed at two separate patient encounters or two separate anatomic locations.

Similarly, if the two corresponding procedures are performed at the same patient encounter and in contiguous structures in the same organ or anatomic region, NCCI PTP-associated modifiers generally shouldn’t be used.

Some Column One/Column Two correct coding edits would never warrant the use of any of the modifiers associated with the NCCI PTP edits. These code pairs are assigned a Correct Coding Modifier Indicator (CCMI) of 0.

NCCI PTP modifier indicators of 9 indicate that the use of NCCI PTP-associated modifiers is not specified. This indicator is used for all code pairs that have a deletion date that is the same as the effective date. This indicator prevents blank spaces from appearing in the indicator field.

Modifier 59 may be used only if no other appropriate modifier describes the service. Claim line edits allow use of NCCI PTP-associated Modifier 91 to bypass them if one or more of the individual laboratory tests are repeated on the same date of service.

The repeat testing must be medically reasonable and necessary and can’t be performed to "confirm initial results; due to testing problems with specimens and equipment or for any other reason when a normal, one-time, reportable result is all that is required."

Procedure to Procedure (PTP) Edits

See also

Examples

search_ptp(hcpcs_code = c("39503", "43116", "33935", "11646"))
#> # A tibble: 158 × 6
#>    hcpcs_code ptp_type      ptp_complements   ptp_deleted ptp_mod ptp_rationale 
#>    <chr>      <chr>         <list>            <date>        <int> <chr>         
#>  1 11646      Comprehensive <tibble [1 × 1]>  9999-12-31        0 Anesthesia se…
#>  2 11646      Comprehensive <tibble [1 × 1]>  2006-12-31        0 Anesthesia se…
#>  3 11646      Comprehensive <tibble [25 × 1]> 9999-12-31        0 Misuse of Col…
#>  4 11646      Comprehensive <tibble [2 × 1]>  2020-12-31        0 Standards of …
#>  5 11646      Comprehensive <tibble [46 × 1]> 9999-12-31        1 Standards of …
#>  6 11646      Comprehensive <tibble [24 × 1]> 9999-12-31        1 Misuse of Col…
#>  7 11646      Comprehensive <tibble [1 × 1]>  2018-12-31        1 CPT Manual or…
#>  8 11646      Comprehensive <tibble [63 × 1]> 9999-12-31        1 CPT Manual or…
#>  9 11646      Comprehensive <tibble [4 × 1]>  9999-12-31        1 Mutually excl…
#> 10 11646      Comprehensive <tibble [11 × 1]> 9999-12-31        0 CPT Manual or…
#> 11 11646      Comprehensive <tibble [1 × 1]>  2015-12-31        1 Misuse of Col…
#> 12 11646      Comprehensive <tibble [1 × 1]>  2016-01-01        9 Misuse of Col…
#> 13 11646      Comprehensive <tibble [2 × 1]>  2016-12-31        0 Standards of …
#> 14 11646      Comprehensive <tibble [5 × 1]>  2016-12-31        0 Misuse of Col…
#> 15 11646      Comprehensive <tibble [26 × 1]> 9999-12-31        0 Standards of …
#> 16 11646      Comprehensive <tibble [3 × 1]>  2019-12-31        0 Standards of …
#> 17 11646      Comprehensive <tibble [1 × 1]>  2015-12-31        0 Standards of …
#> 18 11646      Comprehensive <tibble [2 × 1]>  2009-12-31        0 Misuse of Col…
#> 19 11646      Comprehensive <tibble [1 × 1]>  2018-12-31        0 Standards of …
#> 20 11646      Comprehensive <tibble [1 × 1]>  2009-04-01        9 Standards of …
#> # ℹ 138 more rows

search_ptp(hcpcs_code = "43116",
           ptp_type = "Component",
           ptp_mod = 0)
#> # A tibble: 2 × 6
#>   hcpcs_code ptp_type  ptp_complements  ptp_deleted ptp_mod ptp_rationale       
#>   <chr>      <chr>     <list>           <date>        <int> <chr>               
#> 1 43116      Component <tibble [3 × 1]> 9999-12-31        0 More extensive proc…
#> 2 43116      Component <tibble [1 × 1]> 9999-12-31        0 Mutually exclusive …