Restructured BETOS Classifications

search_rbcs() allows the user to group HCPCS codes into clinically meaningful categories based on the original Berenson-Eggers Type of Service (BETOS) classification.

Usage

search_rbcs(
  hcpcs_code = NULL,
  category = NULL,
  subcategory = NULL,
  family = NULL,
  procedure = NULL,
  concatenate = FALSE,
  ...
)

Arguments

hcpcs_code

<chr> vector of 5-character HCPCS codes

category

<chr> vector of RBCS categories:

Procedure (n = 6920)

Test (n = 3015)

DME (n = 2971)

Treatment (n = 1795)

Imaging (n = 1097)

E&M (n = 695)

Anesthesia (n = 307)

Other (n = 233)

subcategory

<chr> vector of RBCS subcategories (53 unique)

family

<chr> vector of RBCS families (178 unique)

procedure

<chr> Procedure Type:

Major (n = 3676)

Non-Procedure (n = 10113)

Other (n = 3244)

concatenate

<lgl> Concatenate output, default is TRUE

...

These dots are for future extensions and must be empty.

Value

A tibble with the columns:

Column	Description

hcpcs	HCPCS or CPT code

rbcs	RBCS Identifier

rbcs_category	RBCS Category

rbcs_subcategory	RBCS Subcategory

rbcs_family	RBCS Family

rbcs_procedure	RBCS Major Procedure Indicator

date_hcpcs_add	Date HCPCS Code was added

date_hcpcs_end	Date HCPCS Code was no longer effective

date_rbcs_assign	Earliest Date that the RBCS ID was effective

From BETOS to RBCS

The Restructured BETOS Classification System (RBCS) is a taxonomy that allows researchers to group Medicare Part B healthcare service codes into clinically meaningful categories and subcategories.

Based on the original Berenson-Eggers Type of Service (BETOS) classification created in the 1980s, it includes notable updates such as Part B non-physician services and undergoes annual updates by a technical expert panel of researchers and clinicians.

The general framework for grouping service codes into the new RBCS taxonomy largely follows the same structure of BETOS. Like BETOS, the RBCS groups HCPCS codes into categories, subcategories, and families – with categories as the most aggregate level and families as the more granular level.

All Medicare Part B service codes, including non-physician services, are assigned to a 6-character RBCS taxonomy code.

Links

Restructured BETOS Classification System

RBCS Data Dictionary

Update Frequency

Annually

See also

Examples

search_rbcs(hcpcs_code = c("J9264", "39503", "43116", "70170", "0001U"))
#> # A tibble: 6 × 8
#>   hcpcs_code rbcs_procedure rbcs_category rbcs_subcategory           rbcs_family
#>   <chr>      <fct>          <fct>         <chr>                      <chr>      
#> 1 0001U      Non-Procedure  Test          Molecular Testing          No RBCS Fa…
#> 2 0001U      Non-Procedure  Test          Molecular Testing          Genetic An…
#> 3 39503      Major          Procedure     Other Organ Systems        No RBCS Fa…
#> 4 43116      Major          Procedure     Digestive/Gastrointestinal No RBCS Fa…
#> 5 70170      Non-Procedure  Imaging       Standard X-ray             No RBCS Fa…
#> 6 J9264      Non-Procedure  Treatment     Chemotherapy               Chemothera…
#> # ℹ 3 more variables: hcpcs_date_added <date>, hcpcs_date_ended <date>,
#> #   rbcs_date_effective <date>

search_rbcs(hcpcs_code = "0001U")
#> # A tibble: 2 × 8
#>   hcpcs_code rbcs_procedure rbcs_category rbcs_subcategory  rbcs_family     
#>   <chr>      <fct>          <fct>         <chr>             <chr>           
#> 1 0001U      Non-Procedure  Test          Molecular Testing No RBCS Family  
#> 2 0001U      Non-Procedure  Test          Molecular Testing Genetic Analysis
#> # ℹ 3 more variables: hcpcs_date_added <date>, hcpcs_date_ended <date>,
#> #   rbcs_date_effective <date>