HCPCS Action Descriptions

HCPCS Action Descriptions

Usage

switch_action_code(x)

Arguments

x

vector of Action codes

Value

vector of Action descriptions

See also

Examples

dplyr::tibble(code = c(LETTERS),
              desc = switch_action_code(code))
#> # A tibble: 26 × 2
#>    code  desc                                                                   
#>    <chr> <chr>                                                                  
#>  1 A     Added procedure or modifier code                                       
#>  2 B     Change in both administrative data field and long description of proce…
#>  3 C     Change in long description of procedure or modifier code               
#>  4 D     Discontinue procedure or modifier code                                 
#>  5 E     NA                                                                     
#>  6 F     Change in administrative data field of procedure or modifier code      
#>  7 G     NA                                                                     
#>  8 H     NA                                                                     
#>  9 I     NA                                                                     
#> 10 J     NA                                                                     
#> # ℹ 16 more rows