PFS Status Code Descriptions

PFS Status Code Descriptions

Usage

switch_status_description(x)

Arguments

x

vector of status codes

Value

vector of status code descriptions

See also

Examples

dplyr::tibble(
  code = LETTERS,
  name = switch_status_name(LETTERS),
  desc = switch_status_description(LETTERS)
)
#> # A tibble: 26 × 3
#>    code  name                   desc                                            
#>    <chr> <chr>                  <chr>                                           
#>  1 A     Active                 Separately paid if covered. RVUs and payment am…
#>  2 B     Payment Bundle         Payment bundled into payment for other services…
#>  3 C     Carrier Priced         Carriers establish RVUs and payment following d…
#>  4 D     Deleted Codes          Deleted effective with beginning of year.       
#>  5 E     Regulatory Exclusion   Excluded by regulation. No RVUs, no payment mad…
#>  6 F     Deleted/Discontinued   Not subject to 90 day grace period              
#>  7 G     Not Valid for Medicare Another code used for payment. Subject to a 90 …
#>  8 H     Deleted Modifier       Had TC/26 mod in previous year, TC/26 component…
#>  9 I     Not Valid for Medicare Another code used for payment. Not subject to a…
#> 10 J     Anesthesia Service     No RVUs or payment amounts. Only identifies ane…
#> # ℹ 16 more rows