PFS Physician Supervision Descriptions

This field is for use in post payment review.

Usage

switch_supervision(x)

Arguments

x

vector of Physician Supervision indicators

Value

vector of Physician Supervision descriptions

See also

Examples

dplyr::tibble(
  code = c(fuimus::pad_number(
  c(1:6, 9, 21:22, 66, 77)), "6A", "7A"),
  desc = switch_supervision(code))
#> # A tibble: 13 × 2
#>    code  desc                                                                   
#>    <chr> <chr>                                                                  
#>  1 01    Must be performed under General supervision                            
#>  2 02    Must be performed under Direct supervision                             
#>  3 03    Must be performed under Personal supervision                           
#>  4 04    Must be performed under General supervision unless performed by Qualif…
#>  5 05    Must be performed under General supervision unless performed by Qualif…
#>  6 06    Must be performed by Physician or Qualified Electrophysiological Clini…
#>  7 09    Concept does not apply                                                 
#>  8 21    May be performed by Certified Technician under General supervision; ot…
#>  9 22    May be performed by Technician with online real-time Physician contact 
#> 10 66    May be performed by Physician or ABPTS-certified physical therapist wi…
#> 11 77    Must be performed by either ABPTS-certified PT, Uncertified PT under D…
#> 12 6A    In addition to Level 66 rule, ABPTS-certified PT may supervise another…
#> 13 7A    In addition to Level 77 rule, ABPTS-certified PT may supervise another…