Search Claim Adjustment Codes

Search Claim Adjustment Codes

Usage

search_adjustments(adj_code = NULL, adj_type = NULL, ...)

Arguments

adj_code

<chr> vector of Adjustment codes; default is NULL

adj_type

<chr> type of Adjustment code; Group, CARC, RARC; default is NULL

...

These dots are for future extensions and must be empty.

Value

a tibble

Details

Claim Adjustment Reason Codes:

X12 External Code Source 139

These codes describe why a claim or service line was paid differently than it was billed and generally assign responsibility for the adjustment amounts.

The Claim Adjustment Group Codes are internal to the X12 standard. The format is always two alpha characters:

CO (Contractual Obligations): Indicates patient is responsible for remaining balance

CR (Corrections and Reversals)

OA (Other Adjustments): Indicates denial is related to other insurance coverage

PI (Payer-Initiated Reductions)

PR (Patient Responsibility): Indicates patient responsibility has been adjusted

Remittance Advice Remark Codes:

X12 External Code Source 411

Remittance Advice Remark Codes (RARCs) are used to provide additional explanation for an adjustment already described by a Claim Adjustment Reason Code (CARC) or to convey information about remittance processing.

There are two types of RARCs: Supplemental and Informational:

The majority of RARCs are supplemental and, as such, are generally referred to as RARCs without further distinction. Supplemental RARCs provide additional explanation for an adjustment already described by a CARC.

The second type is informational; these are all prefaced with Alert: and are referred to as Alerts. They are used to convey information about remittance processing and are never related to a specific adjustment or CARC.

Examples

search_adjustments(adj_type = "Group")
#> # A tibble: 5 × 8
#>   adj_code adj_type adj_description           adj_usage adj_notes adj_start_date
#>   <chr>    <chr>    <chr>                     <chr>     <chr>     <date>        
#> 1 CO       Group    Contractual Obligations   NA        NA        NA            
#> 2 CR       Group    Corrections and Reversals NA        NA        NA            
#> 3 OA       Group    Other Adjustments         NA        NA        NA            
#> 4 PI       Group    Payer Initiated Reductio… NA        NA        NA            
#> 5 PR       Group    Patient Responsibility    NA        NA        NA            
#> # ℹ 2 more variables: adj_last_modified <date>, adj_end_date <date>

search_adjustments(adj_type = "CARC")
#> # A tibble: 405 × 8
#>    adj_code adj_type adj_description          adj_usage adj_notes adj_start_date
#>    <chr>    <chr>    <chr>                    <chr>     <chr>     <date>        
#>  1 1        CARC     Deductible Amount        NA        NA        1995-01-01    
#>  2 2        CARC     Coinsurance Amount       NA        NA        1995-01-01    
#>  3 3        CARC     Co-payment Amount        NA        NA        1995-01-01    
#>  4 4        CARC     The procedure code is i… Refer to… NA        1995-01-01    
#>  5 5        CARC     The procedure code/type… Refer to… NA        1995-01-01    
#>  6 6        CARC     The procedure/revenue c… Refer to… NA        1995-01-01    
#>  7 7        CARC     The procedure/revenue c… Refer to… NA        1995-01-01    
#>  8 8        CARC     The procedure code is i… Refer to… NA        1995-01-01    
#>  9 9        CARC     The diagnosis is incons… Refer to… NA        1995-01-01    
#> 10 10       CARC     The diagnosis is incons… Refer to… NA        1995-01-01    
#> 11 11       CARC     The diagnosis is incons… Refer to… NA        1995-01-01    
#> 12 12       CARC     The diagnosis is incons… Refer to… NA        1995-01-01    
#> 13 13       CARC     The date of death prece… NA        NA        1995-01-01    
#> 14 14       CARC     The date of birth follo… NA        NA        1995-01-01    
#> 15 15       CARC     The authorization numbe… NA        NA        1995-01-01    
#> 16 16       CARC     Claim/service lacks inf… Do not u… NA        1995-01-01    
#> 17 17       CARC     Requested information w… NA        NA        1995-01-01    
#> 18 18       CARC     Exact duplicate claim/s… NA        NA        1995-01-01    
#> 19 19       CARC     This is a work-related … NA        NA        1995-01-01    
#> 20 20       CARC     This injury/illness is … NA        NA        1995-01-01    
#> # ℹ 385 more rows
#> # ℹ 2 more variables: adj_last_modified <date>, adj_end_date <date>

search_adjustments(adj_type = "RARC")
#> # A tibble: 1,171 × 8
#>    adj_code adj_type adj_description          adj_usage adj_notes adj_start_date
#>    <chr>    <chr>    <chr>                    <chr>     <chr>     <date>        
#>  1 M1       RARC     X-ray not taken within … NA        NA        1997-01-01    
#>  2 M2       RARC     Not paid separately whe… NA        NA        1997-01-01    
#>  3 M3       RARC     Equipment is the same o… NA        NA        1997-01-01    
#>  4 M4       RARC     Alert: This is the last… NA        (Modifie… 1997-01-01    
#>  5 M5       RARC     Monthly rental payments… NA        NA        1997-01-01    
#>  6 M6       RARC     Alert: You must furnish… NA        (Modifie… 1997-01-01    
#>  7 M7       RARC     No rental payments afte… NA        (Modifie… 1997-01-01    
#>  8 M8       RARC     We do not accept blood … NA        NA        1997-01-01    
#>  9 M9       RARC     Alert: This is the tent… NA        (Modifie… 1997-01-01    
#> 10 M10      RARC     Equipment purchases are… NA        NA        1997-01-01    
#> 11 M11      RARC     DME, orthotics and pros… NA        NA        1997-01-01    
#> 12 M12      RARC     Diagnostic tests perfor… NA        NA        1997-01-01    
#> 13 M13      RARC     Only one initial visit … NA        (Modifie… 1997-01-01    
#> 14 M14      RARC     No separate payment for… NA        NA        1997-01-01    
#> 15 M15      RARC     Separately billed servi… NA        NA        1997-01-01    
#> 16 M16      RARC     Alert: Please see our w… NA        (Reactiv… 1997-01-01    
#> 17 M17      RARC     Alert: Payment approved… NA        (Modifie… 1997-01-01    
#> 18 M18      RARC     Certain services may be… NA        (Modifie… 1997-01-01    
#> 19 M19      RARC     Missing oxygen certific… NA        (Modifie… 1997-01-01    
#> 20 M20      RARC     Missing/incomplete/inva… NA        (Modifie… 1997-01-01    
#> # ℹ 1,151 more rows
#> # ℹ 2 more variables: adj_last_modified <date>, adj_end_date <date>