Search Adjustment Codes

CARC and RARC Codes

Usage

search_adjustments(type = NULL, ...)

Arguments

type

<chr> type of Adjustment code; all (default), group, carc, rarc

...

These dots are for future extensions and must be empty.

Value

a tibble

Details

Claim Adjustment Reason Codes:

X12 External Code Source 139

These codes describe why a claim or service line was paid differently than it was billed and generally assign responsibility for the adjustment amounts.

The Claim Adjustment Group Codes are internal to the X12 standard. The format is always two alpha characters:

CO: Contractual Obligations

CR: Corrections and Reversals

OA: Other Adjustments

PI: Payer Initiated Reductions

PR: Patient Responsibility

Remittance Advice Remark Codes:

X12 External Code Source 411

Remittance Advice Remark Codes (RARCs) are used to provide additional explanation for an adjustment already described by a Claim Adjustment Reason Code (CARC) or to convey information about remittance processing.

There are two types of RARCs: Supplemental and Informational:

The majority of RARCs are supplemental and, as such, are generally referred to as RARCs without further distinction. Supplemental RARCs provide additional explanation for an adjustment already described by a CARC.

The second type is informational; these are all prefaced with Alert: and are referred to as Alerts. They are used to convey information about remittance processing and are never related to a specific adjustment or CARC.

Examples

search_adjustments()
#> # A tibble: 1,581 × 8
#>    adj_code adj_type adj_description          adj_usage adj_notes adj_start_date
#>    <chr>    <chr>    <chr>                    <chr>     <chr>     <date>        
#>  1 CO       Group    Contractual Obligations  NA        NA        NA            
#>  2 CR       Group    Corrections and Reversa… NA        NA        NA            
#>  3 OA       Group    Other Adjustments        NA        NA        NA            
#>  4 PI       Group    Payer Initiated Reducti… NA        NA        NA            
#>  5 PR       Group    Patient Responsibility   NA        NA        NA            
#>  6 1        CARC     Deductible Amount        NA        NA        1995-01-01    
#>  7 2        CARC     Coinsurance Amount       NA        NA        1995-01-01    
#>  8 3        CARC     Co-payment Amount        NA        NA        1995-01-01    
#>  9 4        CARC     The procedure code is i… Refer to… NA        1995-01-01    
#> 10 5        CARC     The procedure code/type… Refer to… NA        1995-01-01    
#> 11 6        CARC     The procedure/revenue c… Refer to… NA        1995-01-01    
#> 12 7        CARC     The procedure/revenue c… Refer to… NA        1995-01-01    
#> 13 8        CARC     The procedure code is i… Refer to… NA        1995-01-01    
#> 14 9        CARC     The diagnosis is incons… Refer to… NA        1995-01-01    
#> 15 10       CARC     The diagnosis is incons… Refer to… NA        1995-01-01    
#> 16 11       CARC     The diagnosis is incons… Refer to… NA        1995-01-01    
#> 17 12       CARC     The diagnosis is incons… Refer to… NA        1995-01-01    
#> 18 13       CARC     The date of death prece… NA        NA        1995-01-01    
#> 19 14       CARC     The date of birth follo… NA        NA        1995-01-01    
#> 20 15       CARC     The authorization numbe… NA        NA        1995-01-01    
#> # ℹ 1,561 more rows
#> # ℹ 2 more variables: adj_last_modified <date>, adj_end_date <date>