Level II HCPCS Codes

The Healthcare Common Procedure Coding System (HCPCS) is a collection of codes that represent procedures, supplies, products and services which may be provided to Medicare beneficiaries and to individuals enrolled in private health insurance programs. The codes are divided into two levels, or groups, as described below.

Usage

search_hcpcs(hcpcs_code = NULL, ...)

Arguments

hcpcs_code

<chr> vector of 5-character HCPCS codes

...

These dots are for future extensions and must be empty.

Value

a tibble

Details

Level I Codes and descriptors copyrighted by the American Medical Association's current procedural terminology, fourth edition (CPT-4). These are 5 position numeric codes representing physician and non-physician services.

Level II Includes codes and descriptors copyrighted by the American Dental Association's current dental terminology, (CDT-2023). These are 5 position alpha-numeric codes comprising the d series.

All level II codes and descriptors are approved and maintained jointly by the alpha-numeric editorial panel (consisting of CMS, the Health Insurance Association of America, and the Blue Cross and Blue Shield Association).

These are 5 position alpha-numeric codes representing primarily items and non-physician services that are not represented in the level I codes.

See also

Examples

search_hcpcs(hcpcs_code = c("C9039", "J1835"))
#> # A tibble: 2 × 19
#>   hcpcs_code price multi cov   betos tos   action cim   mcm   statute lab  
#>   <chr>      <chr> <chr> <chr> <chr> <chr> <chr>  <chr> <chr> <chr>   <chr>
#> 1 C9039      53    A     D     O1E   9:F   N      NA    NA    1833(t) NA   
#> 2 J1835      51    A     C     O1E   1:P   N      NA    NA    NA      NA   
#> # ℹ 8 more variables: xref <chr>, asc_grp <chr>, proc <chr>, anes <chr>,
#> #   date_term <date>, date_added <date>, date_asc <date>, date_action <date>