Procedure to Procedure (PTP) Edits

National Correct Coding Initiative (NCCI) Procedure-to-Procedure (PTP) edits prevent inappropriate payment of services that should not be reported together. Each edit has a Column One and Column Two HCPCS/CPT code.

Usage

search_ptp(column_1 = NULL, column_2 = NULL, mod = NULL, ...)

Arguments

column_1

<chr> Comprehensive HCPCS code

column_2

<chr> Component HCPCS code

mod

<int> 1 = Allowed, 0 = Not Allowed, 9 = Not Applicable

...

Empty dots

Value

a tibble

Details

If a provider reports the two codes of an edit pair for the same beneficiary on the same date of service, the Column One code is eligible for payment, but the Column Two code is denied unless a clinically appropriate NCCI PTP-associated modifier is also reported.

PTP Link

Examples

search_ptp(column_1 = c("39503", "43116", "33935", "11646"))
#> # A tibble: 1,345 × 6
#>    column_1 column_2 exist96 deletion   modifier rationale                      
#>    <chr>    <chr>    <chr>   <date>        <int> <chr>                          
#>  1 11646    00170    NA      9999-12-31        0 Anesthesia service included in…
#>  2 11646    01995    NA      2006-12-31        0 Anesthesia service included in…
#>  3 11646    0213T    NA      9999-12-31        0 Misuse of Column Two code with…
#>  4 11646    0216T    NA      9999-12-31        0 Misuse of Column Two code with…
#>  5 11646    0228T    NA      2020-12-31        0 Standards of medical/surgical …
#>  6 11646    0230T    NA      2020-12-31        0 Standards of medical/surgical …
#>  7 11646    0596T    NA      9999-12-31        1 Standards of medical/surgical …
#>  8 11646    0597T    NA      9999-12-31        1 Standards of medical/surgical …
#>  9 11646    11000    NA      9999-12-31        1 Misuse of Column Two code with…
#> 10 11646    11001    NA      9999-12-31        1 Misuse of Column Two code with…
#> # ℹ 1,335 more rows